commit --amend

+++ lib/src/core/use_cases/member/member_crud.dart
@@ -75,6 +75,8 @@ class MemberCrud {
     final user = await authRepo.getCurrentUser();
     if (user == null) throw UserNotFoundError();
     final member = await memberRepo.getMember(id);
+    if (member == null) throw MemberNotFoundError();
+
     final userAsMember = await memberRepo.getCommunityMemberByUserId(
       communityId: member.communityId,
       userId: user.id,
@@ -95,8 +97,10 @@ class MemberCrud {
   Future<void> remove(String memberId) async {
     final user = await authRepo.getCurrentUser();
     if (user != null) {
-      await memberRepo.remove(memberId);
       final member = await memberRepo.getMember(memberId);
+      if (member == null) throw MemberNotFoundError();
+
+      await memberRepo.remove(memberId);
       final userAsMember = await memberRepo.getCommunityMemberByUserId(
           communityId: member.communityId, userId: user.id);
       if (userAsMember != null && userAsMember.role == "member") {
